Move coverage definition into it's own file
